Refactor phone input components with base and pretty variants
