+++ src/components/form/index.ts
@@ -1,7 +1,9 @@
+export * from './BasePhoneInput';
 export * from './DateInput';
 export * from './SSNInput';
 export * from './SelectInput';
 export * from './PhoneInput';
+export * from './PrettyPhoneInput';
 export * from './OTPInput';
 export * from './EmailInput';
 export * from './DateRangeInput';
